add player bottom buttons
